Move autocorrect code to 'unused'
